delete confirmation

+++ app/task-list.php
@@ -51,7 +51,7 @@ require("head.php");
                         <p><?= $taskduedate ?></p>
                         <a href='task-list-details.php?id=<?= $taskid ?>' class='tasklist__taskbutton tasklist__details'></a>
                         <a href='edit-task.php?id=<?= $taskid ?>' class='tasklist__taskbutton tasklist__edit'></a>
-                        <a href="delete-task.php?id=<?= $taskid ?>" class='tasklist__taskbutton tasklist__delete'></a>
+                        <a onclick="return confirm('wirklich Task mit id <?= $taskid ?> löschen?')" href="delete-task.php?id=<?= $taskid ?>" class='tasklist__taskbutton tasklist__delete'></a>
                     </div>
                     <hr>
                 <?php
